% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_scraper.R
\name{holiday_scraper}
\alias{holiday_scraper}
\title{A Holiday Scraper}
\usage{
holiday_scraper(countries = c("sweden", "austria", "norway", "denmark",
  "finland", "germany", "netherlands", "uk", "us"), start_year = 2010,
  end_year = 2020)
}
\arguments{
\item{countries}{Vector with countries for which holidays are wanted. Full name, lower case, with exception of "uk" and "us". Defaults to c("sweden", "norway", "denmark", "finland", "germany", "netherlands", "uk", "us").}

\item{start_year}{Numeric year of which you want the calendar to start from. Defaults to 2010.}

\item{end_year}{Numeric year of which you want the calendar to end at. Defaults to 2020.}
}
\description{
This function allows you to gather holiday dates and their type for specified countries over a specified time period.
Data is gathered from timeanddate.com, so make sure your internet connection is up for it to work properly.
Output data frame "master calendar" with variables:
country - Country for given holiday dates.
year - Year of given holiday dates.
Date - date of the actual holiday listed.
holiday_name - Official name of holiday given in calendar.
holiday_type - Type of holiday. All types are included here, in total 26 different.
Date_fmt - Date output in date format Y-m-d.
}
\examples{
holiday_scraper(countries = c("se", "no", "fi"), start_year = 2015, end_year = 2018)
}
\keyword{calendar}

