% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intra_month.R
\name{intra_month}
\alias{intra_month}
\title{Intra month distribution main function}
\usage{
intra_month(markets = c("at", "de", "dk", "fi", "nl", "no", "se", "uk", "us"),
  min_date = "2017-01-01", max_date = "2017-12-31")
}
\arguments{
\item{markets}{vector containing markets for which distribution is to be estimated.}

\item{min_date}{Start date for month distribution estimate.}

\item{max_date}{End date for month distribution estimate.}
}
\description{
DESCRIPTION HERE
Process using daily_purchases data for historical distribution estimation.
Update this table from time to time. No need to do it for each iteration.
}
\examples{
EXAMPLE HERE
}
\keyword{distribution}
\keyword{intramonth}

