% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intraMonthDist-package.r
\docType{package}
\name{intraMonthDist}
\alias{intraMonthDist}
\alias{intraMonthDist-package}
\title{intraMonthDist.}
\description{
This package creates a daily intramonth volume distribution estimate for Klarna's
existing markets using historical data.
Full year forecast is created using index points adapted to the appearance of holidays and
special days that differ the normal distribution patterns.
}

