% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_by_month.R
\name{normalise_by_month}
\alias{normalise_by_month}
\title{Normalisation function used in the intraMonthDist package.}
\usage{
normalise_by_month(b, nye, f_period)
}
\arguments{
\item{b}{Data frame}

\item{nye}{Extracted information for new years eve}

\item{f_period}{Forecast period}
}
\description{
Further information...
}
\examples{
att <- normalise_by_month(b = .., nye = .., f_period = ..)
}

