% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rd.R
\name{remove_rd}
\alias{remove_rd}
\title{Remove red days; 2 years back in time.}
\usage{
remove_rd(d_p, rd_effect)
}
\arguments{
\item{d_p}{data frame with daily actuals data. Defaults to daily_purchases_act.}

\item{rd_effect}{Matrix with historical red day effect.}
}
\description{
DESCRIPTION HERE
}
\examples{
EXAMPLE HERE
}
\keyword{holiday}
\keyword{red_day}

