% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_rd_2yrsback.R
\name{remove_rd_2yrsback}
\alias{remove_rd_2yrsback}
\title{Remove red days; 2 years back in time.}
\usage{
remove_rd_2yrsback(d_p = daily_purchases_act, rd_effect, markets,
  forecast_range)
}
\arguments{
\item{d_p}{data frame with daily actuals data. Defaults to daily_purchases_act.}

\item{rd_effect}{Empty input matrix for red day effect values.}

\item{markets}{Markets which we want to estimate on.}

\item{forecast_range}{Forecast range.}
}
\description{
DESCRIPTION HERE
}
\examples{
EXAMPLE HERE Change
}
\keyword{holiday}
\keyword{red_day}

